fix: 플레이어 목숨의 잘못된 자료형 수정
float로 되어 있었기 때문에 int로 수정함

+++ Assets/Player/Scripts/ScriptableObjects/PlayerDefaultData.cs
@@ -40,6 +40,6 @@ public class PlayerDefaultData : ScriptableObject
     public float DefaultCurse {get => defaultCurse; set => defaultCurse = value;}
     
     [Header("플레이어 목숨 개수")]
-    [SerializeField] private float defaultLife;
-    public float DefaultLife {get => defaultLife; set => defaultLife = value;}
+    [SerializeField] private int defaultLife;
+    public int DefaultLife {get => defaultLife; set => defaultLife = value;}
 }
